Move data class mixins to models.base
